fix: delete file
