+++ src/features/scan-id-card/components/BoxShadowMask.tsx
