+++ src/features/scan-id-card/components/CaptureButton.tsx
@@ -1,6 +1,5 @@
 import * as React from 'react';
 
-// 1. กำหนด props ที่จะรับ
 interface CaptureButtonProps {
     onClick: () => void;
     isReady: boolean;
@@ -15,8 +14,8 @@ export function CaptureButton({ onClick, isReady }: CaptureButtonProps) {
                   w-14 h-14 rounded-full border-4 border-gray-300 shadow-lg
                   transition-all duration-300
                   ${isReady
-                    ? "bg-white opacity-100 blur-0 cursor-pointer" // สภาพพร้อมถ่าย
-                    : "bg-white/60 cursor-not-allowed"           // สภาพไม่พร้อม
+                    ? "bg-white opacity-100 blur-0 cursor-pointer" 
+                    : "bg-white/60 cursor-not-allowed"          
                 }`}
         />
     );
