+++ src/features/scan-id-card/components/FrameSVG.tsx
