+++ src/features/scan-id-card/components/ScanHeader.tsx
